Plutarch.Test.Bench

Description

Plutarch benchmarking tools

Interface mirrors the one from tasty-bench but bench instead of taking Benchmarkable takes ClosedTerm

To compare benchmark run against baseline file you need to generate it first with cabal run bench -- --csv baseline.csv. Then after making modifications you can rerun the benchmarks to compare with previous values with cabal run bench -- --baseline baseline.csv. You can instruct benchmarks to fail if certain values changed by too much using --fail-if-* flags. See cabal run bench -- --help for all available flags. To regenreate baseline file run with --csv flag again.

Documentation

data BenchConfig

@since WIP

Constructors

Optimizing
Compile with UPLC simplifier pass and no tracing

NonOptimizing Config
Compile without UPLC simplifier and configurable tracing

defaultMain :: TestTree -> IO ()

Use this instead of defaultMain from Test.Tasty to run benchmarks to get formatted output

@since WIP

bench :: TestName -> ClosedTerm a -> TestTree

Create benchmark from Plutarch term without tracing and no UPLC simplifier

@since WIP

benchWithConfig :: TestName -> BenchConfig -> ClosedTerm a -> TestTree

Like bench but with customizable compilation config

@since WIP

bcompare

Arguments

:: String
Tasty pattern to compare as baseline

-> TestTree
Test or test tree to compare with baseline test

-> TestTree

Compare benchmarks, reporting relative CPU, MEM, and size differences

@since WIP

bcompareWithin

Arguments

:: (Double, Double)
CPU bounds

-> (Double, Double)
MEM bounds

-> (Double, Double)
Size bounds

-> String
Tasty pattern to compare as baseline

-> TestTree
Test or test tree to compare with baseline test

-> TestTree

Like bcompare but with customizable upper and lower bounds of relative differences

@since WIP

CLI options

Benchmarks can be compared against CSV file and fail if they differ too much. Run cabal run bench -- --help to see available flags and descriptions. These options are available by default when running benchmarks in defaultMain

data BaselinePath

@since WIP

data CsvPath

@since WIP

data FailIfMoreCpu

@since WIP

data FailIfLessCpu

@since WIP

data FailIfMoreMem

@since WIP

data FailIfLessMem

@since WIP

data FailIfBigger

@since WIP

data FailIfSmaller

@since WIP

Ingredients

These are used by default in defaultMain. You do not need to do anything with them unless you build your own benchmark runner

consoleBenchReporter :: Ingredient

@since WIP

csvReporter :: Ingredient

@since WIP