Plutarch.Test.Golden

Description

Utilities for golden testing

To regenerate golden tests it is enough to remove ./goldens directory and rerun tests

Documentation

data GoldenTestTree

Opaque type representing tree of golden tests

@since WIP

plutarchGolden

Arguments

:: TestName

-> FilePath
Base file name of golden file path.

e.g. "foo" will result in goldens:

./goldens/foo.bench.golden - With execution units and size

./goldens/foo.uplc.eval.golden - With AST after evaluation

./goldens/foo.uplc.golden - With AST before evaluation

-> [GoldenTestTree]

-> TestTree

Convert tree of golden tests into standard Tasty TestTree, capturing results produced by nested golden tests

@since WIP

goldenGroup :: TestName -> [GoldenTestTree] -> GoldenTestTree

Like testGroup but for golden tests

Goldens in the group will be prefixed by the group name

@since WIP

goldenEval :: TestName -> ClosedTerm a -> GoldenTestTree

Like testEval but will append to goldens created by enclosing plutarchGolden

@since WIP

goldenEvalFail :: TestName -> ClosedTerm a -> GoldenTestTree

Like testEvalFail but will append to goldens created by enclosing plutarchGolden

@since WIP